+++ src/Phan/CLI.php
@@ -22,7 +22,7 @@ class CLI
     /**
      * This should be updated to x.y.z-dev after every release, and x.y.z before a release.
      */
-    const PHAN_VERSION = '0.12.2-dev';
+    const PHAN_VERSION = '0.12.2';
 
     /**
      * @var OutputInterface